requirements.txt
whitenoise==5.3.0
gunicorn==20.1.0
Django==4.0.1
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2019.3
pytest-html==3.1.1
coverage==6.3.2
html-testrunner